+++ src/Http/Requests/IndexRequest.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Didasto\RestApi\Http\Requests;
+
+/**
+ * Request der Listenansicht: hier stehen filters(), sortable() und
+ * relations(), keine Validierungsregeln fuer einen Body.
+ */
+class IndexRequest extends RestRequest
+{
+}